changes in the FooterList and added an additional test
