--- components/footer/FooterList2.js
@@ -1,22 +1,0 @@
-export const initiativeLinks = [
-    {
-      label: "About",
-      url: "/about",
-    },
-    {
-      label: "Blog",
-      url: "/blog",
-    },
-    {
-      label: "Shop",
-      url: "https://asyncapi.threadless.com",
-    },
-    {
-      label: "Jobs",
-      url: "/jobs",
-    },
-    {
-      label: "Brand",
-      url: "https://github.com/asyncapi/brand/blob/master/brand-guidelines/README.md",
-    },
-  ];
