+++ cypress/test/footer/footer.cy.js
@@ -1,7 +1,6 @@
 import { mount } from '@cypress/react'
 import Footer from '../../../components/footer/Footer'
-import { socialMediaLinks } from '../../../components/footer/FooterList'
-import { initiativeLinks } from '../../../components/footer/FooterList2'
+import { socialMediaLinks,initiativeLinks} from '../../../components/footer/FooterList'
 
 describe('Footer component', () => {
   beforeEach(() => {
@@ -34,6 +33,10 @@ describe('Footer component', () => {
   it('displays the end content', () => {
     cy.get('[data-testid="Footer-content"]').should('be.visible')
   })
-  
-  
+ 
+  it('checks for logo and link associated with it' , () => {
+    cy.get('[data-testid="Footer-logo-link"]').should('exist')
+    cy.get('[data-testid="Footer-logo-link"]').should('have.attr','href','/')
+
+  })
 })
